Refactor preprocessor of synthesizer to prepare to supprot more datasets
